+++ askbot/conf/access_control.py
@@ -41,6 +41,25 @@ settings.register(
     )
 )
 
+settings.register(
+    livesettings.BooleanValue(
+        ACCESS_CONTROL,
+        'NEW_REGISTRATIONS_DISABLED',
+        default=False,
+        description=_('Disable registration of new users'),
+    )
+)
+
+settings.register(
+    livesettings.LongStringValue(
+        ACCESS_CONTROL,
+        'NEW_REGISTRATIONS_DISABLED_MESSAGE',
+        default=_('<p>New users cannot be registered at this time. Please sign in if you already have an account.</p>'),
+        description=_('Message explaining that user registrations are disabled'),
+        help_text=_('HTML is allowed')
+    )
+)
+
 EMAIL_VALIDATION_CASE_CHOICES = (
     ('nothing', _('nothing - not required')),
     ('see-content', _('account registration')),